requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-filter==25.1
django-heroku==0.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
